:recycle: Refacto MapController
- Moved Geolocation and Directions logic to a dedicated controller
