+++ client/src/components/map/DirectionsController.ts
@@ -1,0 +1,30 @@
+// import MapBoxDirections from '@mapbox/mapbox-gl-directions'
+var MapboxDirections = require('@mapbox/mapbox-gl-directions/dist/mapbox-gl-directions');
+import mapboxgl from 'mapbox-gl'
+import { featureCoords } from '../../utils/map';
+import GeoLocationManager from './GeoLocationController';
+
+class DirectionsManager {
+  public initMapboxDirections = () => {
+
+    const directions = new MapboxDirections({
+      accessToken: mapboxgl.accessToken,
+      unit: 'metric',
+      profile: 'mapbox/walking',
+      controls: false,
+      interactive: false
+    });
+
+    return directions
+  }
+
+  public setFastestPath = (directions: any, markers: any, userLocation: any) => {
+    const nearestMarker = GeoLocationManager.getNearestMarker(markers, userLocation)
+    const nearestMarkerCoords = featureCoords(nearestMarker)
+
+    directions.setOrigin(userLocation);
+    directions.setDestination(nearestMarkerCoords)
+  }
+}
+
+export default new DirectionsManager()
